==> Python_and_R_scripts/1_1_create_GeneCountsMatrix_from_freaturecounts.ipynb <==
#! pip install bioinfokit
# Import of needed packages
import pandas as pd
from tabulate import tabulate
# run this Python code (in a Python interpreter) from a folder where all files are present
from bioinfokit.analys import HtsAna
import re


# make sure all individual count files are present in same folder
# by default, it assumes each count file has .txt extension
HtsAna.merge_featureCount('*.txt', 'Geneid')


from shutil import move
move('gene_matrix_count.csv', 'gene_matrix_count.csv')

df = pd.read_csv("gene_matrix_count.csv", header=0,
                 sep=',', index_col='Geneid')
df


# Define a function to extract the desired substring
def extract_sample_name(column_name):
    pattern = r'([a-zA-Z0-9]+)-([a-zA-Z0-9]+)-([a-zA-Z0-9]+)'
    match = re.search(pattern, column_name)
    sample_name= f'{match.group(2)}-{match.group(3)}'
    return sample_name

# Apply the function to each column and rename the columns
df.columns = df.columns.map(extract_sample_name)


# Display the resulting DataFrame with renamed columns
print(df)

# export the dataframe to a csv file
df.to_csv("gene_matrix_count.csv")

# Create table to sample information 

import pandas as pd


# Extract conditions from column names
conditions = []
for column in df.columns:
    if 'R' in column:
        # Use regex to find the number after 'R'
        match = re.search(r'R(\d+)', column)
        if match:
            number_after_r = int(match.group(1))
            if number_after_r >= 20:
                conditions.append('reoxygenation24h')
            else:
                conditions.append('reoxygenation3h')
    elif 'A' in column:
        conditions.append('anoxia')
    elif 'N' in column:
        conditions.append('normoxia')
    else:
        conditions.append('unknown')  # or handle other cases

# Create a new DataFrame
sample_info = pd.DataFrame({'Geneid': df.columns, 'condition': conditions[:len(df.columns)]})

# Display the resulting DataFrame
print(sample_info)

# Export to csv
sample_info.to_csv('sample_info_3hr.csv', index=False)
